vertexnormal and pixel normal extraction works again

+++ com.unity.render-pipelines.universal/Shaders/LitForwardPass.hlsl
@@ -129,7 +129,7 @@ Varyings LitPassVertex(Attributes input)
 
 
 // Used in Standard (Physically Based) shader
-float4 LitPassFragment(Varyings input) : SV_Target
+half4 LitPassFragment(Varyings input) : SV_Target
 {
     UNITY_SETUP_INSTANCE_ID(input);
     UNITY_SETUP_STEREO_EYE_INDEX_POST_VERTEX(input);
